requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.1
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
